backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
Werkzeug==2.3.6
gunicorn==21.2.0
